refactored pick and place example
