+++ drake/examples/kuka_iiwa_arm/dev/pick_and_place/BUILD
@@ -16,18 +16,22 @@ drake_cc_library(
     name = "kuka_pick_and_place",
     srcs = [
         "action.cc",
+        "pick_and_place_common.cc",
         "world_state.cc",
     ],
     hdrs = [
         "action.h",
+        "pick_and_place_common.h",
         "world_state.h",
     ],
     deps = [
+        "//drake/common/trajectories:piecewise_quaternion",
+        "//drake/examples/kuka_iiwa_arm/dev:iiwa_ik_planner",
         "//drake/examples/kuka_iiwa_arm:iiwa_common",
         "//drake/lcmtypes:schunk",
+        "//drake/util:lcm_util",
         "//drake/multibody:rigid_body_tree",
         "//drake/multibody/parsers",
-        "//drake/util:lcm_util",
     ],
 )
 
@@ -38,11 +42,11 @@ drake_cc_binary(
     ],
     deps = [
         ":kuka_pick_and_place",
-        "//drake/common/trajectories:piecewise_polynomial",
-        "//drake/common/trajectories:piecewise_quaternion",
+        "//drake/examples/kuka_iiwa_arm/dev:iiwa_ik_planner",
         "//drake/examples/kuka_iiwa_arm:iiwa_common",
         "//drake/examples/kuka_iiwa_arm:iiwa_lcm",
-        "//drake/examples/kuka_iiwa_arm/dev:iiwa_ik_planner",
+        "//drake/common/trajectories:piecewise_quaternion",
+        "//drake/common/trajectories:piecewise_polynomial",
         "//drake/examples/schunk_wsg:schunk_wsg_lcm",
         "//drake/util:lcm_util",
     ],
